move figure generation in build
