+++ athena/entrypoints/visualize.py
@@ -6,8 +6,10 @@ from athena.core.config import DataConfig, IndicatorsConfig
 from athena.core.context import ProjectContext
 from athena.core.interfaces import DatasetLayout, Fluctuations
 from athena.entrypoints.utils import load_config
-from athena.tradingtools.indicators.build import build_indicator_lines
-from athena.tradingtools.indicators.chart import build_and_save_indicators_figure
+from athena.tradingtools.indicators.build import (
+    build_indicator_lines,
+    build_and_save_indicators_figure,
+)
 
 
 @click.command()
